We now support additional passes
If we cannot solve the order in the first pass, we now run additional passes on the sublist that could not be resolved recursive until we cannot solve the sublist. If we cannot resolve the pass, it means that the ordering of the votes are final, and that two votes will be places on the same place in the final result.